+++ PCIT_core/internal/llvm_interface/include/class_impls/native_ptr_decls.h
@@ -30,6 +30,7 @@ namespace llvm{
 	class FunctionType;
 	class IntegerType;
 	class PointerType;
+	class ArrayType;
 	class StructType;
 
 	// stmts