+++ PIR/include/ReaderAgent.h
@@ -33,9 +33,14 @@ namespace pcit::pir{
 			ReaderAgent(const Module& _module, const Function& func) : module(_module), target_func(&func) {}
 			~ReaderAgent() = default;
 
+			EVO_NODISCARD auto getModule() const -> const Module& { return this->module; }
 			EVO_NODISCARD auto hasTargetFunction() const -> bool { return this->target_func != nullptr; }
-			EVO_NODISCARD auto setTargetFunction(const Function& func) -> void { this->target_func = &func; }
-			EVO_NODISCARD auto clearTargetFunction() -> void { this->target_func = nullptr; }
+			auto setTargetFunction(const Function& func) -> void { this->target_func = &func; }
+			auto clearTargetFunction() -> void { this->target_func = nullptr; }
+			EVO_NODISCARD auto getTargetFunction() const -> const Function& {
+				evo::debugAssert(this->hasTargetFunction(), "ReaderAgent has no target function set");
+				return *this->target_func;
+			}
 
 
 			EVO_NODISCARD auto getExprType(const Expr& expr) const -> Type;
@@ -55,6 +60,7 @@ namespace pcit::pir{
 			EVO_NODISCARD auto getAlloca(const Expr& expr) const -> const Alloca&;
 			EVO_NODISCARD auto getLoad(const Expr& expr) const -> const Load&;
 			EVO_NODISCARD auto getStore(const Expr& expr) const -> const Store&;
+			EVO_NODISCARD auto getCalcPtr(const Expr& expr) const -> const CalcPtr&;
 
 			EVO_NODISCARD auto getAdd(const Expr& expr) const -> const Add&;
 			EVO_NODISCARD auto getAddWrap(const Expr& expr) const -> const AddWrap&;